+++ frontend/src/app/app-routing.module.ts
@@ -1,12 +1,12 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {RouterModule, Routes} from "@angular/router";
-import {EntryComponent} from "./components/entry/entry.component";
-import {ProfileComponent} from "./components/profile/profile.component";
+import {EntryComponent} from "./modules/entry/entry.component";
+import {ProfileComponent} from "./modules/profile/profile.component";
 import {AuthGuard} from "./guards/auth.guard";
-import {ExperimentalComponent} from "./components/experimental/experimental.component";
-import {LobbyComponent} from "./components/lobby/lobby.component";
-import {WerewolfLobbyComponent} from "./components/lobby/werewolf-lobby/werewolf-lobby.component";
+import {ExperimentalComponent} from "./modules/experimental/experimental.component";
+import {LobbyComponent} from "./modules/lobby/lobby.component";
+import {WerewolfLobbyComponent} from "./modules/lobby/werewolf-lobby/werewolf-lobby.component";
 
 const routes: Routes = [
   {path: '', redirectTo: 'entry', pathMatch: 'full'},