+++ frontend/src/app/app.module.ts
@@ -2,23 +2,23 @@ import {NgModule, OnInit} from '@angular/core';
 import {BrowserModule} from '@angular/platform-browser';
 
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
-import {EntryComponent} from './components/entry/entry.component';
+import {EntryComponent} from './modules/entry/entry.component';
 import {FormsModule, ReactiveFormsModule} from "@angular/forms";
 import {HttpClientModule} from "@angular/common/http";
 
 import {authInterceptorProviders} from './interceptors/auth.interceptor';
 import {AppRoutingModule} from './app-routing.module';
 import {AppComponent} from "./app.component";
-import {DialogContentSignupDialog, SignupComponent} from './components/authentication/signup/signup.component';
+import {DialogContentSignupDialog, SignupComponent} from './modules/authentication/signup/signup.component';
 import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
 import {ToastrModule} from "ngx-toastr";
-import {DialogContentSigninDialog, SigninComponent} from './components/authentication/signin/signin.component';
+import {DialogContentSigninDialog, SigninComponent} from './modules/authentication/signin/signin.component';
 import {
   DialogContentProfileEditDialog,
   DialogContentProfileEnterGameDialog,
   ProfileComponent
-} from './components/profile/profile.component';
-import {LoadingIndicatorComponent} from './components/utility/loading-indicator/loading-indicator.component';
+} from './modules/profile/profile.component';
+import {LoadingIndicatorComponent} from './modules/utility/loading-indicator/loading-indicator.component';
 import {MatChipsModule} from "@angular/material/chips";
 import {MatCardModule} from "@angular/material/card";
 import {MatButtonModule} from "@angular/material/button";
@@ -30,11 +30,11 @@ import {MatProgressSpinnerModule} from "@angular/material/progress-spinner";
 import {MatAutocompleteModule} from "@angular/material/autocomplete";
 import {MatOptionModule} from "@angular/material/core";
 import {MatIconModule} from "@angular/material/icon";
-import { ExperimentalComponent } from './components/experimental/experimental.component';
+import { ExperimentalComponent } from './modules/experimental/experimental.component';
 import {OverlayContainer} from "@angular/cdk/overlay";
 import {MatMenuModule} from "@angular/material/menu";
-import { LobbyComponent } from './components/lobby/lobby.component';
-import { WerewolfLobbyComponent } from './components/lobby/werewolf-lobby/werewolf-lobby.component';
+import { LobbyComponent } from './modules/lobby/lobby.component';
+import { WerewolfLobbyComponent } from './modules/lobby/werewolf-lobby/werewolf-lobby.component';
 import {DragDropModule} from "@angular/cdk/drag-drop";
 import {MatListModule} from "@angular/material/list";
 import {MatProgressBarModule} from "@angular/material/progress-bar";